requirements.txt
feedparser
newspaper3k
python-dotenv
openai>=1.14
markdown